% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-code-chunks-v3.R
\name{add_keys}
\alias{add_keys}
\title{Add RDB unique keys to a build table function.}
\usage{
add_keys(filename, rdb.keys)
}
\description{
Add a set of meta-data variables contained in the \code{rdbkeys} file and
unique keys needed for filtering or merging tables generated from the
concordance files.
}
\details{
This is an internal function only.
}
