% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FULL_DATABASE.R
\name{bind_data}
\alias{bind_data}
\title{Combine Batched CSV Files}
\usage{
bind_data(years)
}
\arguments{
\item{years}{A vector of tax years to compile. These correspond to subdirectories containing raw table files.}
}
\value{
Saves compiled data tables as CSV and RDS files in the "COMPILED" directory.
}
\description{
Combines all raw table files generated during parallel processing into compiled data.
}
\details{
The function loops through raw table files, consolidates them by table name, and ensures unique entries. Results are stored for each tax year.
}
\examples{
\dontrun{
bind_data(years = 2015:2017)
}
}
