% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODE_CHUNKS.R
\name{build_all_chunks}
\alias{build_all_chunks}
\title{Build all table functions from the concordance file.}
\usage{
build_all_chunks()
}
\description{
Recreate all table build functions (each table is a chunk) fresh from the concordance.
}
