% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-F9-P00-T00-HEADER.R
\name{BUILD_F9_P00_T00_HEADER}
\alias{BUILD_F9_P00_T00_HEADER}
\title{Build table F9-P00-T00-HEADER}
\usage{
BUILD_F9_P00_T00_HEADER(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
