% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-F9-P01-T00-SUMMARY.R
\name{BUILD_F9_P01_T00_SUMMARY}
\alias{BUILD_F9_P01_T00_SUMMARY}
\title{Build table F9-P01-T00-SUMMARY}
\usage{
BUILD_F9_P01_T00_SUMMARY(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
