% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_F9_P01_T00_SUMMARY_EZ.R
\name{BUILD_F9_P01_T00_SUMMARY_EZ}
\alias{BUILD_F9_P01_T00_SUMMARY_EZ}
\title{Build table F9-P01-T00-SUMMARY-EZ}
\usage{
BUILD_F9_P01_T00_SUMMARY_EZ(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
