% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-F9-P02-T00-SIGNATURE.R
\name{BUILD_F9_P02_T00_SIGNATURE}
\alias{BUILD_F9_P02_T00_SIGNATURE}
\title{Build table F9-P02-T00-SIGNATURE}
\usage{
BUILD_F9_P02_T00_SIGNATURE(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
