% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_F9_P03_T00_MISSION.R
\name{BUILD_F9_P03_T00_MISSION}
\alias{BUILD_F9_P03_T00_MISSION}
\title{Build table F9-P03-T00-MISSION}
\usage{
BUILD_F9_P03_T00_MISSION(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
