% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_F9_P03_T00_PROGRAM_ONE.R
\name{BUILD_F9_P03_T00_PROGRAM_ONE}
\alias{BUILD_F9_P03_T00_PROGRAM_ONE}
\title{Build table F9-P03-T00-PROGRAM-ONE}
\usage{
BUILD_F9_P03_T00_PROGRAM_ONE(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
