% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_F9_P03_T00_PROGRAM_THREE.R
\name{BUILD_F9_P03_T00_PROGRAM_THREE}
\alias{BUILD_F9_P03_T00_PROGRAM_THREE}
\title{Build table F9-P03-T00-PROGRAM-THREE}
\usage{
BUILD_F9_P03_T00_PROGRAM_THREE(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
