% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_F9_P03_T00_PROGRAM_TWO.R
\name{BUILD_F9_P03_T00_PROGRAM_TWO}
\alias{BUILD_F9_P03_T00_PROGRAM_TWO}
\title{Build table F9-P03-T00-PROGRAM-TWO}
\usage{
BUILD_F9_P03_T00_PROGRAM_TWO(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
