% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-F9-P03-T00-PROGRAMS.R
\name{BUILD_F9_P03_T00_PROGRAMS}
\alias{BUILD_F9_P03_T00_PROGRAMS}
\title{Build table F9-P03-T00-PROGRAMS}
\usage{
BUILD_F9_P03_T00_PROGRAMS(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
