% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_F9_P03_T01_PROGRAMS_OTHER.R
\name{BUILD_F9_P03_T01_PROGRAMS_OTHER}
\alias{BUILD_F9_P03_T01_PROGRAMS_OTHER}
\title{Build table F9-P03-T01-PROGRAMS-OTHER}
\usage{
BUILD_F9_P03_T01_PROGRAMS_OTHER(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
