% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_F9_P03_T02_PROGRAMS_EZ.R
\name{BUILD_F9_P03_T02_PROGRAMS_EZ}
\alias{BUILD_F9_P03_T02_PROGRAMS_EZ}
\title{Build table F9-P03-T02-PROGRAMS-EZ}
\usage{
BUILD_F9_P03_T02_PROGRAMS_EZ(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
