% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-F9-P04-T00-REQUIRED-SCHEDULES.R
\name{BUILD_F9_P04_T00_REQUIRED_SCHEDULES}
\alias{BUILD_F9_P04_T00_REQUIRED_SCHEDULES}
\title{Build table F9-P04-T00-REQUIRED-SCHEDULES}
\usage{
BUILD_F9_P04_T00_REQUIRED_SCHEDULES(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
