% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_F9_P05_T00_OTHER_IRS_FILING.R
\name{BUILD_F9_P05_T00_OTHER_IRS_FILING}
\alias{BUILD_F9_P05_T00_OTHER_IRS_FILING}
\title{Build table F9-P05-T00-OTHER-IRS-FILING}
\usage{
BUILD_F9_P05_T00_OTHER_IRS_FILING(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
