% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-F9-P06-T00-GOVERNANCE.R
\name{BUILD_F9_P06_T00_GOVERNANCE}
\alias{BUILD_F9_P06_T00_GOVERNANCE}
\title{Build table F9-P06-T00-GOVERNANCE}
\usage{
BUILD_F9_P06_T00_GOVERNANCE(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
