% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_F9_P06_T00_GOVERNANCE_EZ.R
\name{BUILD_F9_P06_T00_GOVERNANCE_EZ}
\alias{BUILD_F9_P06_T00_GOVERNANCE_EZ}
\title{Build table F9-P06-T00-GOVERNANCE-EZ}
\usage{
BUILD_F9_P06_T00_GOVERNANCE_EZ(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
