% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-F9-P07-T00-DIR-TRUST-KEY.R
\name{BUILD_F9_P07_T00_DIR_TRUST_KEY}
\alias{BUILD_F9_P07_T00_DIR_TRUST_KEY}
\title{Build table F9-P07-T00-DIR-TRUST-KEY}
\usage{
BUILD_F9_P07_T00_DIR_TRUST_KEY(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
