% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_F9_P07_T01_COMPENSATION.R
\name{BUILD_F9_P07_T01_COMPENSATION}
\alias{BUILD_F9_P07_T01_COMPENSATION}
\title{Build table F9-P07-T01-COMPENSATION}
\usage{
BUILD_F9_P07_T01_COMPENSATION(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
