% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_F9_P07_T01_COMPENSATION_HCE_EZ.R
\name{BUILD_F9_P07_T01_COMPENSATION_HCE_EZ}
\alias{BUILD_F9_P07_T01_COMPENSATION_HCE_EZ}
\title{Build table F9-P07-T01-COMPENSATION-HCE-EZ}
\usage{
BUILD_F9_P07_T01_COMPENSATION_HCE_EZ(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
