% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_F9_P07_T02_CONTRACTORS.R
\name{BUILD_F9_P07_T02_CONTRACTORS}
\alias{BUILD_F9_P07_T02_CONTRACTORS}
\title{Build table F9-P07-T02-CONTRACTORS}
\usage{
BUILD_F9_P07_T02_CONTRACTORS(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
