% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-F9-P08-T00-REVENUE.R
\name{BUILD_F9_P08_T00_REVENUE}
\alias{BUILD_F9_P08_T00_REVENUE}
\title{Build table F9-P08-T00-REVENUE}
\usage{
BUILD_F9_P08_T00_REVENUE(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
