% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_F9_P08_T01_REVENUE_PROGRAMS.R
\name{BUILD_F9_P08_T01_REVENUE_PROGRAMS}
\alias{BUILD_F9_P08_T01_REVENUE_PROGRAMS}
\title{Build table F9-P08-T01-REVENUE-PROGRAMS}
\usage{
BUILD_F9_P08_T01_REVENUE_PROGRAMS(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
