% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_F9_P08_T02_REVENUE_MISC.R
\name{BUILD_F9_P08_T02_REVENUE_MISC}
\alias{BUILD_F9_P08_T02_REVENUE_MISC}
\title{Build table F9-P08-T02-REVENUE-MISC}
\usage{
BUILD_F9_P08_T02_REVENUE_MISC(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
