% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-F9-P09-T00-EXPENSES.R
\name{BUILD_F9_P09_T00_EXPENSES}
\alias{BUILD_F9_P09_T00_EXPENSES}
\title{Build table F9-P09-T00-EXPENSES}
\usage{
BUILD_F9_P09_T00_EXPENSES(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
