% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_F9_P09_T01_EXPENSES_OTHER.R
\name{BUILD_F9_P09_T01_EXPENSES_OTHER}
\alias{BUILD_F9_P09_T01_EXPENSES_OTHER}
\title{Build table F9-P09-T01-EXPENSES-OTHER}
\usage{
BUILD_F9_P09_T01_EXPENSES_OTHER(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
