% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-F9-P10-T00-BALANCE-SHEET.R
\name{BUILD_F9_P10_T00_BALANCE_SHEET}
\alias{BUILD_F9_P10_T00_BALANCE_SHEET}
\title{Build table F9-P10-T00-BALANCE-SHEET}
\usage{
BUILD_F9_P10_T00_BALANCE_SHEET(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
