% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_F9_P11_T00_ASSETS.R
\name{BUILD_F9_P11_T00_ASSETS}
\alias{BUILD_F9_P11_T00_ASSETS}
\title{Build table F9-P11-T00-ASSETS}
\usage{
BUILD_F9_P11_T00_ASSETS(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
