% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_F9_P12_T00_FINANCIAL_REPORTING.R
\name{BUILD_F9_P12_T00_FINANCIAL_REPORTING}
\alias{BUILD_F9_P12_T00_FINANCIAL_REPORTING}
\title{Build table F9-P12-T00-FINANCIAL-REPORTING}
\usage{
BUILD_F9_P12_T00_FINANCIAL_REPORTING(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
