% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-index.R
\name{build_index}
\alias{build_index}
\title{Build Efiler Index}
\usage{
build_index(tax.years = 2009:2020)
}
\arguments{
\item{file.years}{Numeric or character vector of years starting in 2011.}
}
\value{
Returns a data frame of the AWS index file.

See the tinyindex for a full list of variables in
the index file.
}
\description{
Build an index of IRS 990 efilers available on AWS
}
\details{
AWS has reference index files in JSON format available
at \url{https://s3.amazonaws.com/irs-form-990/index_YEAR} where
YEAR is the year of interest. Note that these index files
are organized by filing dates, not the tax year for the
nonprofit. Unless the fiscal year ends in December the tax
year is a year before the filing date (for example, individuals
file 2019 income taxes in April of 2020).
}
\examples{
\dontrun{
index <- build_index( years=2015:2018 )
}
}
