% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FULL_DATABASE.R
\name{build_one_year}
\alias{build_one_year}
\title{Database Build Steps for a Year of Data}
\usage{
build_one_year(year, num.cores)
}
\arguments{
\item{year}{The year of data you would like processed.}
}
\value{
The build_tables() function serves as the workhorse
when parsing XML files into rectangular RDB tables. The
function saves batched CSV files in the YEAR subfolder.
}
\description{
Using the BATCHFILE created by create_batchfiles(),
load the list of batched URLs, then send each
batch to a CORE for XML parsing using parallel
processing functions.
}
\details{
The function should be run from the main project
folder that contains the YEAR subfolders created by
create_batchfiles(). A YEARX folder containing a BATCHFILE
should exist when calling build_one_year( year=YEARX ).
}
\examples{
create_batchfiles( tinyindex, years=2020:2022, group.size=100 )
build_one_year( year=2021, index=tinyindex )
}
