% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb-build-function-adapted.R
\name{build_rdb_table_adapted}
\alias{build_rdb_table_adapted}
\title{build_rdb_table_adapted (improved from rdb-functions-v2)}
\usage{
build_rdb_table_adapted(url, table.name)
}
\description{
rdb build function equipped to handle tables with sub-grouping
it is generalized so it still works on previously working tables
}
