% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR-P01-T01-rdb-build-function.R
\name{build_rdb_table_SR_P01_T01}
\alias{build_rdb_table_SR_P01_T01}
\title{some title text}
\usage{
build_rdb_table_SR_P01_T01(url)
}
\description{
some description text
}
\details{
some additional details
}
