% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SA_P00_T00_HEADER.R
\name{BUILD_SA_P00_T00_HEADER}
\alias{BUILD_SA_P00_T00_HEADER}
\title{Build table SA-P00-T00-HEADER}
\usage{
BUILD_SA_P00_T00_HEADER(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
