% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SA_P01_T01_PUBLIC_CHARITY_STATUS.R
\name{BUILD_SA_P01_T01_PUBLIC_CHARITY_STATUS}
\alias{BUILD_SA_P01_T01_PUBLIC_CHARITY_STATUS}
\title{Build table SA-P01-T01-PUBLIC-CHARITY-STATUS}
\usage{
BUILD_SA_P01_T01_PUBLIC_CHARITY_STATUS(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
