% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-SA-P02-T00-SUPPORT_SCHEDULE_170.R
\name{BUILD_SA_P02_T00_SUPPORT_SCHEDULE_170}
\alias{BUILD_SA_P02_T00_SUPPORT_SCHEDULE_170}
\title{Build table SA-P02-T00-SUPPORT_SCHEDULE_170}
\usage{
BUILD_SA_P02_T00_SUPPORT_SCHEDULE_170(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
