% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SA_P03_T00_SUPPORT_SCHEDULE_509.R
\name{BUILD_SA_P03_T00_SUPPORT_SCHEDULE_509}
\alias{BUILD_SA_P03_T00_SUPPORT_SCHEDULE_509}
\title{Build table SA-P03-T00-SUPPORT_SCHEDULE_509}
\usage{
BUILD_SA_P03_T00_SUPPORT_SCHEDULE_509(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
