% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-SA-P04-T00-SUPPORT-ORGS.R
\name{BUILD_SA_P04_T00_SUPPORT_ORGS}
\alias{BUILD_SA_P04_T00_SUPPORT_ORGS}
\title{Build table SA-P04-T00-SUPPORT-ORGS}
\usage{
BUILD_SA_P04_T00_SUPPORT_ORGS(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
