% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-SA-P99-T00-PUBLIC-CHARITY-STATUS.R
\name{BUILD_SA_P99_T00_PUBLIC_CHARITY_STATUS}
\alias{BUILD_SA_P99_T00_PUBLIC_CHARITY_STATUS}
\title{Build table SA-P99-T00-PUBLIC-CHARITY-STATUS}
\usage{
BUILD_SA_P99_T00_PUBLIC_CHARITY_STATUS(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
