% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SB_P01_T01_CONTRIBUTORS.R
\name{BUILD_SB_P01_T01_CONTRIBUTORS}
\alias{BUILD_SB_P01_T01_CONTRIBUTORS}
\title{Build table SB-P01-T01-CONTRIBUTORS}
\usage{
BUILD_SB_P01_T01_CONTRIBUTORS(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
