% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SC_P01_T01_POLITICAL_ORGS_INFO.R
\name{BUILD_SC_P01_T01_POLITICAL_ORGS_INFO}
\alias{BUILD_SC_P01_T01_POLITICAL_ORGS_INFO}
\title{Build table SC-P01-T01-POLITICAL-ORGS-INFO}
\usage{
BUILD_SC_P01_T01_POLITICAL_ORGS_INFO(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
