% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-SC-P02-T00-LOBBY.R
\name{BUILD_SC_P02_T00_LOBBY}
\alias{BUILD_SC_P02_T00_LOBBY}
\title{Build table SC-P02-T00-LOBBY}
\usage{
BUILD_SC_P02_T00_LOBBY(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
