% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SC_P04_T99_SUPPLEMENTAL_INFO.R
\name{BUILD_SC_P04_T99_SUPPLEMENTAL_INFO}
\alias{BUILD_SC_P04_T99_SUPPLEMENTAL_INFO}
\title{Build table SC-P04-T99-SUPPLEMENTAL-INFO}
\usage{
BUILD_SC_P04_T99_SUPPLEMENTAL_INFO(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
