% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SXX_TABLE.R
\name{BUILD_SCHEDULE_TABLE}
\alias{BUILD_SCHEDULE_TABLE}
\title{Retrieve list of filed schedules}
\usage{
BUILD_SCHEDULE_TABLE(doc, url)
}
\arguments{
\item{doc}{A parsed XML document}

\item{url}{AWS location of the XML doc}
}
\value{
Returns a data frame with a table recording
the presence of each schedule in the tax filing.
}
\description{
Check whether a specific organization has included any schedules with their return.
}
\details{
Create an index so that it is possible to identify which organizations
have filed which schedule each year.
}
\examples{
\dontrun{
url <- tinyindex$URL[1]
doc <- xml2::read_xml( url )
xml2::xml_ns_strip( doc ) 
BUILD_SCHEDULE_TABLE( doc, url )
}
}
