% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SD_P01_T00_ORGS_DONOR_ADVISED_FUNDS_OTH.R
\name{BUILD_SD_P01_T00_ORGS_DONOR_ADVISED_FUNDS_OTH}
\alias{BUILD_SD_P01_T00_ORGS_DONOR_ADVISED_FUNDS_OTH}
\title{Build table SD-P01-T00-ORGS-DONOR-ADVISED-FUNDS-OTH}
\usage{
BUILD_SD_P01_T00_ORGS_DONOR_ADVISED_FUNDS_OTH(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
