% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SD_P02_T00_CONSERV_EASEMENTS.R
\name{BUILD_SD_P02_T00_CONSERV_EASEMENTS}
\alias{BUILD_SD_P02_T00_CONSERV_EASEMENTS}
\title{Build table SD-P02-T00-CONSERV-EASEMENTS}
\usage{
BUILD_SD_P02_T00_CONSERV_EASEMENTS(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
