% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/BUILD_SD_P03_T00_ORGS_COLLECT_ART_HIST_TREASURE_OTH.R
\name{BUILD_SD_P03_T00_ORGS_COLLECT_ART_HIST_TREASURE_OTH}
\alias{BUILD_SD_P03_T00_ORGS_COLLECT_ART_HIST_TREASURE_OTH}
\title{Build table SD-P03-T00-ORGS-COLLECT-ART-HIST-TREASURE-OTH}
\usage{
BUILD_SD_P03_T00_ORGS_COLLECT_ART_HIST_TREASURE_OTH(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
