% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/BUILD_SD_P04_T00_ESCROW_CUSTODIAL_ARRANGEMENTS.R
\name{BUILD_SD_P04_T00_ESCROW_CUSTODIAL_ARRANGEMENTS}
\alias{BUILD_SD_P04_T00_ESCROW_CUSTODIAL_ARRANGEMENTS}
\title{Build table SD-P04-T00-ESCROW-CUSTODIAL-ARRANGEMENTS}
\usage{
BUILD_SD_P04_T00_ESCROW_CUSTODIAL_ARRANGEMENTS(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
