% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SD_P05_T00_ENDOWMENT.R
\name{BUILD_SD_P05_T00_ENDOWMENT}
\alias{BUILD_SD_P05_T00_ENDOWMENT}
\title{Build table SD-P05-T00-ENDOWMENT}
\usage{
BUILD_SD_P05_T00_ENDOWMENT(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
