% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-SD-P06-T00-LAND-BLDG-EQUIP.R
\name{BUILD_SD_P06_T00_LAND_BLDG_EQUIP}
\alias{BUILD_SD_P06_T00_LAND_BLDG_EQUIP}
\title{Build table SD-P06-T00-LAND-BLDG-EQUIP}
\usage{
BUILD_SD_P06_T00_LAND_BLDG_EQUIP(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
