% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-SD-P07-T00-INVESTMENTS-OTH-SECURITIES.R
\name{BUILD_SD_P07_T00_INVESTMENTS_OTH_SECURITIES}
\alias{BUILD_SD_P07_T00_INVESTMENTS_OTH_SECURITIES}
\title{Build table SD-P07-T00-INVESTMENTS-OTH-SECURITIES}
\usage{
BUILD_SD_P07_T00_INVESTMENTS_OTH_SECURITIES(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
