% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SD_P07_T00_INVESTMENTS_SECURITIES.R
\name{BUILD_SD_P07_T00_INVESTMENTS_SECURITIES}
\alias{BUILD_SD_P07_T00_INVESTMENTS_SECURITIES}
\title{Build table SD-P07-T00-INVESTMENTS-SECURITIES}
\usage{
BUILD_SD_P07_T00_INVESTMENTS_SECURITIES(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
