% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SD_P07_T01_INVESTMENTS_OTH_DERIVATIVES.R
\name{BUILD_SD_P07_T01_INVESTMENTS_OTH_DERIVATIVES}
\alias{BUILD_SD_P07_T01_INVESTMENTS_OTH_DERIVATIVES}
\title{Build table SD-P07-T01-INVESTMENTS-OTH-DERIVATIVES}
\usage{
BUILD_SD_P07_T01_INVESTMENTS_OTH_DERIVATIVES(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
