% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SD_P07_T01_INVESTMENTS_OTH_EQUITY.R
\name{BUILD_SD_P07_T01_INVESTMENTS_OTH_EQUITY}
\alias{BUILD_SD_P07_T01_INVESTMENTS_OTH_EQUITY}
\title{Build table SD-P07-T01-INVESTMENTS-OTH-EQUITY}
\usage{
BUILD_SD_P07_T01_INVESTMENTS_OTH_EQUITY(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
