% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SD_P07_T01_INVESTMENTS_OTH_SECURITIES.R
\name{BUILD_SD_P07_T01_INVESTMENTS_OTH_SECURITIES}
\alias{BUILD_SD_P07_T01_INVESTMENTS_OTH_SECURITIES}
\title{Build table SD-P07-T01-INVESTMENTS-OTH-SECURITIES}
\usage{
BUILD_SD_P07_T01_INVESTMENTS_OTH_SECURITIES(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
