% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SD_P08_T01_INVESTMENTS_PROG_RLTD.R
\name{BUILD_SD_P08_T01_INVESTMENTS_PROG_RLTD}
\alias{BUILD_SD_P08_T01_INVESTMENTS_PROG_RLTD}
\title{Build table SD-P08-T01-INVESTMENTS-PROG-RLTD}
\usage{
BUILD_SD_P08_T01_INVESTMENTS_PROG_RLTD(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
