% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SD_P09_T00_OTH_ASSETS.R
\name{BUILD_SD_P09_T00_OTH_ASSETS}
\alias{BUILD_SD_P09_T00_OTH_ASSETS}
\title{Build table SD-P09-T00-OTH-ASSETS}
\usage{
BUILD_SD_P09_T00_OTH_ASSETS(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
