% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-SD-P10-T00-OTH-LIABILITIES.R
\name{BUILD_SD_P10_T00_OTH_LIABILITIES}
\alias{BUILD_SD_P10_T00_OTH_LIABILITIES}
\title{Build table SD-P10-T00-OTH-LIABILITIES}
\usage{
BUILD_SD_P10_T00_OTH_LIABILITIES(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
