% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SD_P10_T01_OTH_LIABILITIES.R
\name{BUILD_SD_P10_T01_OTH_LIABILITIES}
\alias{BUILD_SD_P10_T01_OTH_LIABILITIES}
\title{Build table SD-P10-T01-OTH-LIABILITIES}
\usage{
BUILD_SD_P10_T01_OTH_LIABILITIES(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
