% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-SD-P11-T00-RECONCILIATION-REVENUE.R
\name{BUILD_SD_P11_T00_RECONCILIATION_REVENUE}
\alias{BUILD_SD_P11_T00_RECONCILIATION_REVENUE}
\title{Build table SD-P11-T00-RECONCILIATION-REVENUE}
\usage{
BUILD_SD_P11_T00_RECONCILIATION_REVENUE(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
