% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SD_P12_T00_RECONCILIATION_EXPENSES.R
\name{BUILD_SD_P12_T00_RECONCILIATION_EXPENSES}
\alias{BUILD_SD_P12_T00_RECONCILIATION_EXPENSES}
\title{Build table SD-P12-T00-RECONCILIATION-EXPENSES}
\usage{
BUILD_SD_P12_T00_RECONCILIATION_EXPENSES(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
