% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SD_P99_T00_RECONCILIATION_NETASSETS.R
\name{BUILD_SD_P99_T00_RECONCILIATION_NETASSETS}
\alias{BUILD_SD_P99_T00_RECONCILIATION_NETASSETS}
\title{Build table SD-P99-T00-RECONCILIATION-NETASSETS}
\usage{
BUILD_SD_P99_T00_RECONCILIATION_NETASSETS(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
