% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-SE-P01-T00-SCHOOLS.R
\name{BUILD_SE_P01_T00_SCHOOLS}
\alias{BUILD_SE_P01_T00_SCHOOLS}
\title{Build table SE-P01-T00-SCHOOLS}
\usage{
BUILD_SE_P01_T00_SCHOOLS(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
