% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SE_P02_T99_SUPPLEMENTAL_INFO.R
\name{BUILD_SE_P02_T99_SUPPLEMENTAL_INFO}
\alias{BUILD_SE_P02_T99_SUPPLEMENTAL_INFO}
\title{Build table SE-P02-T99-SUPPLEMENTAL-INFO}
\usage{
BUILD_SE_P02_T99_SUPPLEMENTAL_INFO(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
