% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SF_P01_T00_FRGN_ACTS.R
\name{BUILD_SF_P01_T00_FRGN_ACTS}
\alias{BUILD_SF_P01_T00_FRGN_ACTS}
\title{Build table SF-P01-T00-FRGN-ACTS}
\usage{
BUILD_SF_P01_T00_FRGN_ACTS(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
