% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SF_P01_T01_FRGN_ACTS_BY_REGION.R
\name{BUILD_SF_P01_T01_FRGN_ACTS_BY_REGION}
\alias{BUILD_SF_P01_T01_FRGN_ACTS_BY_REGION}
\title{Build table SF-P01-T01-FRGN-ACTS-BY-REGION}
\usage{
BUILD_SF_P01_T01_FRGN_ACTS_BY_REGION(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
