% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SF_P02_T01_FRGN_ORG_GRANTS.R
\name{BUILD_SF_P02_T01_FRGN_ORG_GRANTS}
\alias{BUILD_SF_P02_T01_FRGN_ORG_GRANTS}
\title{Build table SF-P02-T01-FRGN-ORG-GRANTS}
\usage{
BUILD_SF_P02_T01_FRGN_ORG_GRANTS(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
