% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SF_P03_T01_FRGN_INDIV_GRANTS.R
\name{BUILD_SF_P03_T01_FRGN_INDIV_GRANTS}
\alias{BUILD_SF_P03_T01_FRGN_INDIV_GRANTS}
\title{Build table SF-P03-T01-FRGN-INDIV-GRANTS}
\usage{
BUILD_SF_P03_T01_FRGN_INDIV_GRANTS(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
