% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-SF-P04-T00-FRGN-INTERESTS.R
\name{BUILD_SF_P04_T00_FRGN_INTERESTS}
\alias{BUILD_SF_P04_T00_FRGN_INTERESTS}
\title{Build table SF-P04-T00-FRGN-INTERESTS}
\usage{
BUILD_SF_P04_T00_FRGN_INTERESTS(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
