% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SF_P05_T99_EXPLANATION_TEXT.R
\name{BUILD_SF_P05_T99_EXPLANATION_TEXT}
\alias{BUILD_SF_P05_T99_EXPLANATION_TEXT}
\title{Build table SF-P05-T99-EXPLANATION-TEXT}
\usage{
BUILD_SF_P05_T99_EXPLANATION_TEXT(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
