% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-SF-P99-T00-FRGN-ORG-GRANTS.R
\name{BUILD_SF_P99_T00_FRGN_ORG_GRANTS}
\alias{BUILD_SF_P99_T00_FRGN_ORG_GRANTS}
\title{Build table SF-P99-T00-FRGN-ORG-GRANTS}
\usage{
BUILD_SF_P99_T00_FRGN_ORG_GRANTS(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
