% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SG_P01_T00_FUNDRAISING_ACTS.R
\name{BUILD_SG_P01_T00_FUNDRAISING_ACTS}
\alias{BUILD_SG_P01_T00_FUNDRAISING_ACTS}
\title{Build table SG-P01-T00-FUNDRAISING-ACTS}
\usage{
BUILD_SG_P01_T00_FUNDRAISING_ACTS(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
