% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SG_P01_T01_FUNDRAISERS_INFO.R
\name{BUILD_SG_P01_T01_FUNDRAISERS_INFO}
\alias{BUILD_SG_P01_T01_FUNDRAISERS_INFO}
\title{Build table SG-P01-T01-FUNDRAISERS-INFO}
\usage{
BUILD_SG_P01_T01_FUNDRAISERS_INFO(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
