% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-SG-P02-T00-FUNDRAISING-EVENTS.R
\name{BUILD_SG_P02_T00_FUNDRAISING_EVENTS}
\alias{BUILD_SG_P02_T00_FUNDRAISING_EVENTS}
\title{Build table SG-P02-T00-FUNDRAISING-EVENTS}
\usage{
BUILD_SG_P02_T00_FUNDRAISING_EVENTS(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
