% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SG_P02_T01_FUNDRAISING_EVENTS.R
\name{BUILD_SG_P02_T01_FUNDRAISING_EVENTS}
\alias{BUILD_SG_P02_T01_FUNDRAISING_EVENTS}
\title{Build table SG-P02-T01-FUNDRAISING-EVENTS}
\usage{
BUILD_SG_P02_T01_FUNDRAISING_EVENTS(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
