% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-SG-P03-T00-GAMING.R
\name{BUILD_SG_P03_T00_GAMING}
\alias{BUILD_SG_P03_T00_GAMING}
\title{Build table SG-P03-T00-GAMING}
\usage{
BUILD_SG_P03_T00_GAMING(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
