% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SH_P01_T00_FAP_COMMUNITY_BENEFIT_POLICY.R
\name{BUILD_SH_P01_T00_FAP_COMMUNITY_BENEFIT_POLICY}
\alias{BUILD_SH_P01_T00_FAP_COMMUNITY_BENEFIT_POLICY}
\title{Build table SH-P01-T00-FAP-COMMUNITY-BENEFIT-POLICY}
\usage{
BUILD_SH_P01_T00_FAP_COMMUNITY_BENEFIT_POLICY(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
