% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SH_P04_T01_COMPANY_JOINT_VENTURES.R
\name{BUILD_SH_P04_T01_COMPANY_JOINT_VENTURES}
\alias{BUILD_SH_P04_T01_COMPANY_JOINT_VENTURES}
\title{Build table SH-P04-T01-COMPANY-JOINT-VENTURES}
\usage{
BUILD_SH_P04_T01_COMPANY_JOINT_VENTURES(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
