% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SH_P05_T01_HOSPITAL_FACILITY.R
\name{BUILD_SH_P05_T01_HOSPITAL_FACILITY}
\alias{BUILD_SH_P05_T01_HOSPITAL_FACILITY}
\title{Build table SH-P05-T01-HOSPITAL-FACILITY}
\usage{
BUILD_SH_P05_T01_HOSPITAL_FACILITY(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
