% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-SI-P01-T00-GRANTS-INFO.R
\name{BUILD_SI_P01_T00_GRANTS_INFO}
\alias{BUILD_SI_P01_T00_GRANTS_INFO}
\title{Build table SI-P01-T00-GRANTS-INFO}
\usage{
BUILD_SI_P01_T00_GRANTS_INFO(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
