% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SI_P02_T01_GRANTS_US_ORGS_GOVTS.R
\name{BUILD_SI_P02_T01_GRANTS_US_ORGS_GOVTS}
\alias{BUILD_SI_P02_T01_GRANTS_US_ORGS_GOVTS}
\title{Build table SI-P02-T01-GRANTS-US-ORGS-GOVTS}
\usage{
BUILD_SI_P02_T01_GRANTS_US_ORGS_GOVTS(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
