% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SI_P03_T01_GRANTS_US_INDIV.R
\name{BUILD_SI_P03_T01_GRANTS_US_INDIV}
\alias{BUILD_SI_P03_T01_GRANTS_US_INDIV}
\title{Build table SI-P03-T01-GRANTS-US-INDIV}
\usage{
BUILD_SI_P03_T01_GRANTS_US_INDIV(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
