% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-SI-P99-T00-GRANTS-US-ORGS-GOVTS.R
\name{BUILD_SI_P99_T00_GRANTS_US_ORGS_GOVTS}
\alias{BUILD_SI_P99_T00_GRANTS_US_ORGS_GOVTS}
\title{Build table SI-P99-T00-GRANTS-US-ORGS-GOVTS}
\usage{
BUILD_SI_P99_T00_GRANTS_US_ORGS_GOVTS(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
