% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-SJ-P01-T00-COMPENSATION.R
\name{BUILD_SJ_P01_T00_COMPENSATION}
\alias{BUILD_SJ_P01_T00_COMPENSATION}
\title{Build table SJ-P01-T00-COMPENSATION}
\usage{
BUILD_SJ_P01_T00_COMPENSATION(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
