% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SJ_P02_T01_COMPENSATION_DTK.R
\name{BUILD_SJ_P02_T01_COMPENSATION_DTK}
\alias{BUILD_SJ_P02_T01_COMPENSATION_DTK}
\title{Build table SJ-P02-T01-COMPENSATION-DTK}
\usage{
BUILD_SJ_P02_T01_COMPENSATION_DTK(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
