% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SK_P01_T01_BOND_ISSUES.R
\name{BUILD_SK_P01_T01_BOND_ISSUES}
\alias{BUILD_SK_P01_T01_BOND_ISSUES}
\title{Build table SK-P01-T01-BOND-ISSUES}
\usage{
BUILD_SK_P01_T01_BOND_ISSUES(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
