% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SK_P02_T01_BOND_PROCEEDS.R
\name{BUILD_SK_P02_T01_BOND_PROCEEDS}
\alias{BUILD_SK_P02_T01_BOND_PROCEEDS}
\title{Build table SK-P02-T01-BOND-PROCEEDS}
\usage{
BUILD_SK_P02_T01_BOND_PROCEEDS(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
