% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SK_P03_T01_BOND_PRIVATE_BIZ_USE.R
\name{BUILD_SK_P03_T01_BOND_PRIVATE_BIZ_USE}
\alias{BUILD_SK_P03_T01_BOND_PRIVATE_BIZ_USE}
\title{Build table SK-P03-T01-BOND-PRIVATE-BIZ-USE}
\usage{
BUILD_SK_P03_T01_BOND_PRIVATE_BIZ_USE(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
