% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SK_P04_T01_BOND_ARBITRAGE.R
\name{BUILD_SK_P04_T01_BOND_ARBITRAGE}
\alias{BUILD_SK_P04_T01_BOND_ARBITRAGE}
\title{Build table SK-P04-T01-BOND-ARBITRAGE}
\usage{
BUILD_SK_P04_T01_BOND_ARBITRAGE(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
