% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SK_P05_T01_PROCEDURE_CORRECTIVE_ACT.R
\name{BUILD_SK_P05_T01_PROCEDURE_CORRECTIVE_ACT}
\alias{BUILD_SK_P05_T01_PROCEDURE_CORRECTIVE_ACT}
\title{Build table SK-P05-T01-PROCEDURE-CORRECTIVE-ACT}
\usage{
BUILD_SK_P05_T01_PROCEDURE_CORRECTIVE_ACT(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
