% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SL_P01_T01_EXCESS_BENEFIT_TRANSAC.R
\name{BUILD_SL_P01_T01_EXCESS_BENEFIT_TRANSAC}
\alias{BUILD_SL_P01_T01_EXCESS_BENEFIT_TRANSAC}
\title{Build table SL-P01-T01-EXCESS-BENEFIT-TRANSAC}
\usage{
BUILD_SL_P01_T01_EXCESS_BENEFIT_TRANSAC(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
