% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-SL-P02-T00-LOANS-INTERESTED-PERS.R
\name{BUILD_SL_P02_T00_LOANS_INTERESTED_PERS}
\alias{BUILD_SL_P02_T00_LOANS_INTERESTED_PERS}
\title{Build table SL-P02-T00-LOANS-INTERESTED-PERS}
\usage{
BUILD_SL_P02_T00_LOANS_INTERESTED_PERS(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
