% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SL_P03_T01_GRANTS_INTERESTED_PERS.R
\name{BUILD_SL_P03_T01_GRANTS_INTERESTED_PERS}
\alias{BUILD_SL_P03_T01_GRANTS_INTERESTED_PERS}
\title{Build table SL-P03-T01-GRANTS-INTERESTED-PERS}
\usage{
BUILD_SL_P03_T01_GRANTS_INTERESTED_PERS(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
