% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SL_P04_T01_BIZ_TRANSAC_INTERESTED_PERS.R
\name{BUILD_SL_P04_T01_BIZ_TRANSAC_INTERESTED_PERS}
\alias{BUILD_SL_P04_T01_BIZ_TRANSAC_INTERESTED_PERS}
\title{Build table SL-P04-T01-BIZ-TRANSAC-INTERESTED-PERS}
\usage{
BUILD_SL_P04_T01_BIZ_TRANSAC_INTERESTED_PERS(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
