% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHUNKS-SM-P01-T00-NONCASH-CONTRIBUTIONS.R
\name{BUILD_SM_P01_T00_NONCASH_CONTRIBUTIONS}
\alias{BUILD_SM_P01_T00_NONCASH_CONTRIBUTIONS}
\title{Build table SM-P01-T00-NONCASH-CONTRIBUTIONS}
\usage{
BUILD_SM_P01_T00_NONCASH_CONTRIBUTIONS(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
