% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SM_P01_T01_NONCASH_CONTRIBUTIONS.R
\name{BUILD_SM_P01_T01_NONCASH_CONTRIBUTIONS}
\alias{BUILD_SM_P01_T01_NONCASH_CONTRIBUTIONS}
\title{Build table SM-P01-T01-NONCASH-CONTRIBUTIONS}
\usage{
BUILD_SM_P01_T01_NONCASH_CONTRIBUTIONS(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
