% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/CHUNKS-SN-P01-T00-LIQUIDATION-TERMINATION-DISSOLUTION.R
\name{BUILD_SN_P01_T00_LIQUIDATION_TERMINATION_DISSOLUTION}
\alias{BUILD_SN_P01_T00_LIQUIDATION_TERMINATION_DISSOLUTION}
\title{Build table SN-P01-T00-LIQUIDATION-TERMINATION-DISSOLUTION}
\usage{
BUILD_SN_P01_T00_LIQUIDATION_TERMINATION_DISSOLUTION(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
