% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SN_P02_T00_DISPOSITION_OF_ASSETS.R
\name{BUILD_SN_P02_T00_DISPOSITION_OF_ASSETS}
\alias{BUILD_SN_P02_T00_DISPOSITION_OF_ASSETS}
\title{Build table SN-P02-T00-DISPOSITION-OF-ASSETS}
\usage{
BUILD_SN_P02_T00_DISPOSITION_OF_ASSETS(doc, url)
}
\description{
Generate a 1:1 table for the relational database.
}
