% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SN_P03_T99_SUPPLEMENTAL_INFO.R
\name{BUILD_SN_P03_T99_SUPPLEMENTAL_INFO}
\alias{BUILD_SN_P03_T99_SUPPLEMENTAL_INFO}
\title{Build table SN-P03-T99-SUPPLEMENTAL-INFO}
\usage{
BUILD_SN_P03_T99_SUPPLEMENTAL_INFO(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
