% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SR_P01_T01_ID_DISREGARDED_ENTITIES.R
\name{BUILD_SR_P01_T01_ID_DISREGARDED_ENTITIES}
\alias{BUILD_SR_P01_T01_ID_DISREGARDED_ENTITIES}
\title{Build table SR-P01-T01-ID-DISREGARDED-ENTITIES}
\usage{
BUILD_SR_P01_T01_ID_DISREGARDED_ENTITIES(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
