% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SR_P02_T01_ID_RLTD_TAX_EXEMPED_ORGS.R
\name{BUILD_SR_P02_T01_ID_RLTD_TAX_EXEMPED_ORGS}
\alias{BUILD_SR_P02_T01_ID_RLTD_TAX_EXEMPED_ORGS}
\title{Build table SR-P02-T01-ID-RLTD-TAX-EXEMPED-ORGS}
\usage{
BUILD_SR_P02_T01_ID_RLTD_TAX_EXEMPED_ORGS(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
