% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/BUILD_SR_P03_T01_ID_RLTD_ORGS_TAXABLE_PARTNERSHIP.R
\name{BUILD_SR_P03_T01_ID_RLTD_ORGS_TAXABLE_PARTNERSHIP}
\alias{BUILD_SR_P03_T01_ID_RLTD_ORGS_TAXABLE_PARTNERSHIP}
\title{Build table SR-P03-T01-ID-RLTD-ORGS-TAXABLE-PARTNERSHIP}
\usage{
BUILD_SR_P03_T01_ID_RLTD_ORGS_TAXABLE_PARTNERSHIP(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
