% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_SR_P05_T01_TRANSACTIONS_RLTD_ORGS.R
\name{BUILD_SR_P05_T01_TRANSACTIONS_RLTD_ORGS}
\alias{BUILD_SR_P05_T01_TRANSACTIONS_RLTD_ORGS}
\title{Build table SR-P05-T01-TRANSACTIONS-RLTD-ORGS}
\usage{
BUILD_SR_P05_T01_TRANSACTIONS_RLTD_ORGS(doc, url)
}
\description{
Generate a 1:M table for the relational database.
}
