% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_FUNCTIONS.R
\name{build_tables}
\alias{build_tables}
\title{Build Tables}
\usage{
build_tables(urls, year, fx.names = NULL, table.names = NULL)
}
\arguments{
\item{urls}{A vector of URLs to process.}

\item{year}{The tax year associated with the data.}

\item{fx.names}{A vector of function names for processing tables.}

\item{table.names}{Optional vector of table names. Defaults to NULL.}
}
\value{
A vector of failed URLs.
}
\description{
Extracts and writes table data from URLs.
}
\examples{
failed_urls <- build_tables(urls, year = 2023)
}
