% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-one-to-one-tables.R
\name{build_tables_parallel}
\alias{build_tables_parallel}
\title{some title text}
\usage{
build_tables_parallel(groups, cores, year, table.names = NULL)
}
\description{
some description text
}
\details{
some additional details
}
