% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb-build-function-adapted.R
\name{collapse_nodes}
\alias{collapse_nodes}
\title{collapse_nodes (new)}
\usage{
collapse_nodes(lc.xpaths)
}
\description{
helper function:
takes in a list of possible roots and finds sub-roots if existing
i.e. collapses possible subparenting/subgrouping structure in xml files
}
