% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concordance}
\alias{concordance}
\title{The Master Concordance File that serves as a crosswalk between xpaths and variable names.}
\format{
A data frame with 6782 rows and 22 variables:
\describe{
\item{xpath}{the xml document path for a variable}
\item{variable_name}{the primary variable label for the dataset}
\item{description}{}
\item{location_code_xsd}{}
\item{location_code_family}{}
\item{location_code}{}
\item{form}{}
\item{form_type}{}
\item{form_part}{}
\item{form_line_number}{}
\item{variable_scope}{}
\item{data_type_xsd}{}
\item{data_type_simple}{}
\item{rdb_relationship}{}
\item{rdb_table}{}
\item{required}{}
\item{versions}{}
\item{latest_version}{}
\item{duplicated}{}
\item{current_version}{}
\item{production_rule}{}
\item{validated}{}
}
}
\source{
\url{https://github.com/Nonprofit-Open-Data-Collective/irs-efile-master-concordance-file}
}
\description{
The concordance file is the Rosetta Stone of the IRS efile database.
It documents the type of mapping needed to convert XML files into
rectangular relational database tables appropriate for research.
}
\details{
There are multiple versions of 990 forms (the 990 and 990EZ)
and multiple versions of each form over time, so each variable is
is associated with multiple xpaths describing its possible location
in an xml document. The primary purpose of the \code{concordance} file is
to organize xpaths by variable, and to organize variables into tables,
as well as provide documentatoin used to generate the data dictionary.
}
