% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{count_xpaths}
\alias{count_xpaths}
\title{Aggregate Missing XPATHS}
\usage{
count_xpaths(df)
}
\arguments{
\item{df}{A data frame of missing xpaths created by get_missing_xpath_df()}
}
\description{
Read a MISSING-XPATHS data frame from get_missing_xpath_df()
and return a table with counts of occurrences of each xpath
along with the list of unique xsd versions the correspond to the xpath.
}
\examples{
df <- get_missing_xpath_df( "2020" )
count_xpaths( df )
}
