% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FULL_DATABASE.R
\name{create_batchfiles}
\alias{create_batchfiles}
\title{Apply split_index() to All Years}
\usage{
create_batchfiles(index, years, group.size)
}
\arguments{
\item{index}{An index data frame containing columns \code{URL} and \code{TaxYear}.}

\item{years}{A vector of years that will be included in the build.}

\item{group.size}{An integer specifying the size of each batch used in parallel processing.}
}
\value{
NULL.
}
\description{
Create YEAR folders and one BATCHFILE list in each.
}
\examples{
create_batchfiles( tinyindex, years=2020:2022, group.size=100 )
}
