% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODE_CHUNKS.R
\name{create_code_chunk}
\alias{create_code_chunk}
\title{Concordance to table functions}
\usage{
create_code_chunk(rdb.table = "F9-P00-T00-HEADER", show = FALSE)
}
\description{
Turn the concordance file into a set of functions to build 1-to-1 tables.
}
\details{
Note that this function only works for RDB tables with one-to-one cardinality
(see the \code{rdb_relationship} variable in the \code{concordance} file for cardinality).
}
\examples{
\dontrun{
  data( concordance )
  # list of all one-to-one table
  unique( concordance$rdb_tables[ concordance$rdb_relationship == "ONE" ] ) 
  # create a new function to build a single table
  create_code_chunks( rdb.table="F9-P00-T00-HEADER", show=TRUE )
  # create functions for all one-to-one tables 
  do_all_chunks() 
}

}
