% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIAGRAM_TABLE_STRUCTURE.R
\name{create_edgelist_v2}
\alias{create_edgelist_v2}
\title{Create Edgelist Version 2}
\usage{
create_edgelist_v2(xpaths)
}
\arguments{
\item{xpaths}{A character vector of XPath expressions.}
}
\value{
A data frame containing the edgelist with two columns representing parent-child relationships.
}
\description{
Similar to create_edgelist_v1 but removes the prefixes "Return/ReturnData" from XPath expressions.
}
\examples{
create_edgelist_v2(c("/Return/ReturnData/A/B", "/Return/ReturnData/C/D"))
}
