% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-code-chunks-v3.R
\name{do_all_chunks}
\alias{do_all_chunks}
\title{Create all one-to-one tables.}
\usage{
do_all_chunks()
}
\description{
Convert the concordance into a set of table build functions
for all of the one-to-one tables in the concordance file.
}
