% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_FUNCTIONS_RDB.R
\name{drop_empty_nodes}
\alias{drop_empty_nodes}
\title{Drop empty XML nodesets}
\usage{
drop_empty_nodes(nodeset)
}
\arguments{
\item{nodeset}{An xml2 xml_nodeset.}
}
\value{
Subset of non-empty nodes as xml2 xml_nodeset.
}
\description{
Drop empty XML nodesets
}
\examples{
x <- 
"<doc> 
   <RevenueGrp> <BusinessCd> 611430 </BusinessCd> </RevenueGrp> 
   <RevenueGrp> <BusinessCd> 611431 </BusinessCd> </RevenueGrp>  
   <RevenueGrp/>               # void nodeset
   <RevenueGrp></RevenueGrp>   # empty nodeset 
 </doc>"
doc <- xml2::read_xml(x)
nodes <- xml2::xml_find_all( doc, ".//RevenueGrp" )
nodes \%>\% xml2::xml_text()  # empty nodes ""
drop_empty_nodes( nodes )
}
