% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIST_FUNCTIONS_BY_FILE.R
\name{extract_functions}
\alias{extract_functions}
\title{Extract Function Names and Arguments from an R Script}
\usage{
extract_functions(file)
}
\arguments{
\item{file}{A character string specifying the path to an R script file.}
}
\value{
A data frame containing function names and their respective arguments.
}
\description{
Parses an R script to identify function definitions and their arguments.
}
\examples{
extract_functions("script.R")
}
