% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb-build-function-adapted.R
\name{find_group_names_2}
\alias{find_group_names_2}
\title{find_group_names_2 (improved from rdb-functions-v2)}
\usage{
find_group_names_2(table.name)
}
\description{
identifies all possible roots/xpaths from a table name
}
