% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_AWS_INDEX.R
\name{find_most_recent_date}
\alias{find_most_recent_date}
\title{Find Most Recent Date}
\usage{
find_most_recent_date(dates)
}
\arguments{
\item{dates}{A character vector of dates in "YYYY-MM-DD" format.}
}
\value{
An integer indicating the position of the most recent date.
}
\description{
Finds the position of the most recent date in a vector of dates.
}
\examples{
dates <- c( "2023-11-18", "2023-11-19", "2024-06-01" )
find_most_recent_date( dates )
}
