% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_FUNCTIONS_RDB.R
\name{find_table_headers}
\alias{find_table_headers}
\title{Find Table Headers}
\usage{
find_table_headers(table.name)
}
\arguments{
\item{table.name}{A character string representing the name of the table for which headers are to be found.}
}
\value{
A character vector of consolidated and updated group names for the table headers.
}
\description{
Identifies and consolidates table headers for a specified database table.
}
\details{
This function retrieves and processes XPaths from a concordance table to identify group names associated with a table's headers.
}
\examples{
headers <- find_table_headers("F9-P03-T01-PROGRAMS")
print(headers)
}
