% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{format_ein}
\alias{format_ein}
\title{Format EIN}
\usage{
format_ein(x, to = "id")
}
\arguments{
\item{x}{A vector of EINs.}

\item{to}{Specify output format: "id" for EIN-XX-XXXXXX or "n" for numeric only.}
}
\value{
A vector of formatted EINs.
}
\description{
Formats Employer Identification Numbers (EINs) into a standard format.
}
\examples{
formatted_ein <- format_ein(c("12345678"), to = "id")
}
