% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIST_FUNCTIONS_BY_FILE.R
\name{generate_ascii_diagram}
\alias{generate_ascii_diagram}
\title{Generate an ASCII Diagram of Functions and Arguments from R Scripts}
\usage{
generate_ascii_diagram(files)
}
\arguments{
\item{files}{A character vector of R script file paths.}
}
\value{
Prints an ASCII diagram showing the functions and their arguments.
}
\description{
Reads multiple R script files and generates an ASCII representation
listing each file, its contained functions, and their arguments.
}
\examples{
generate_ascii_diagram(c("script1.R", "script2.R"))
}
