% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_AWS_INDEX.R
\name{get_all_batch_indices_awscli}
\alias{get_all_batch_indices_awscli}
\title{Get All Batch Index Filenames Using CLI}
\usage{
get_all_batch_indices_awscli()
}
\value{
A character string representing the filename of the most recent index.
}
\description{
Retrieves the filename of the most recent AWS index.
}
\examples{
\dontrun{
get_all_batch_indices_awscli()
}
}
