% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FULL_DATABASE.R
\name{get_batch_ids}
\alias{get_batch_ids}
\title{Get Batch Labels from a BATCHFILE}
\usage{
get_batch_ids(path = ".")
}
\arguments{
\item{path}{The location of the BATCHFILE, defaults to the current directory.}
}
\value{
A character vector of batch IDs.
}
\description{
Batch ids contain the batch name + size (number of URLs in the batch)
using the format NAME{SIZE} (g1{100}, g2{100}, etc.).
}
\examples{
create_batchfiles( tinyindex, years=2020:2022, group.size=100 )
get_batch_ids( path=2021 )
}
