% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FULL_DATABASE.R
\name{get_batch_names}
\alias{get_batch_names}
\title{Parse Batch IDs and Return Batch Names}
\usage{
get_batch_names(batch.ids)
}
\arguments{
\item{batch.ids}{Vector created by get_batch_ids()}
}
\value{
A character vector of batch names.
}
\description{
Parse the batch IDs (e.g. g1{100},g2{70}) and return
only the batch names ("g1","g2"). The batch names
are used to retrieve URLs from the BATCHFILE and
to delete the group from BATCHFILE after processing.
}
\examples{
batch.ids <- c("g1{100}","g2{100}","g3{50}")
get_batch_names( batch.ids )
}
