% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FULL_DATABASE.R
\name{get_batchfile}
\alias{get_batchfile}
\title{Retrieve a Batchfile for a Given Year}
\usage{
get_batchfile(year)
}
\arguments{
\item{year}{Specifies which batchfile to retrieve.}
}
\value{
batchfile data frame.
}
\description{
Load a batchfile (URLs split into groups) generated from create_batchfiles().
}
\examples{
create_batchfiles( tinyindex, years=2020, group.size=100 )
bf <- get_batchfile( 2020 )
names( bf )  # group IDs
}
