% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{get_concordance}
\alias{get_concordance}
\title{Retrieve the Concordance File}
\usage{
get_concordance(gh = TRUE)
}
\arguments{
\item{gh}{Logical; if TRUE, fetches the concordance file from GitHub. If FALSE, uses a local dataset.}
}
\value{
A data.table containing the concordance data.
}
\description{
Downloads the master concordance file from GitHub or loads a local version.
}
\examples{
concordance <- get_concordance(gh = TRUE)
head(concordance)
}
