% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_AWS_INDEX.R
\name{get_current_index_batch}
\alias{get_current_index_batch}
\title{Load Most Recent Data Commons Batch Index}
\usage{
get_current_index_batch(TIMEOUT = 600)
}
\value{
A data.table containing the most recent batch
index file (only records new to this batch).
}
\description{
Loads the most recent batch index file into a data.table.
}
\examples{
\dontrun{
df <- get_current_index_batch()
}
}
