% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_AWS_INDEX.R
\name{get_current_index_full_awscli}
\alias{get_current_index_full_awscli}
\title{Get Current AWS Index Using CLI}
\usage{
get_current_index_full_awscli()
}
\value{
A character string representing the filename of the most recent index.
}
\description{
Retrieves the filename of the most recent AWS index.
}
\examples{
\dontrun{
get_current_index_full_awscli()
}
}
