% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_FUNCTIONS.R
\name{get_fx_names}
\alias{get_fx_names}
\title{Get Function Names}
\usage{
get_fx_names(table.names = NULL, exclude = c("T99"))
}
\arguments{
\item{table.names}{An optional vector of table names. Defaults to all tables in the concordance.}

\item{exclude}{A vector of substrings to exclude from table names.}
}
\value{
A vector of function names.
}
\description{
Retrieves a list of function names based on table names and exclusion criteria.
}
\examples{
fx_names <- get_fx_names(table.names = c("TABLE1", "TABLE2"))
}
