% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_FUNCTIONS.R
\name{get_fxdf}
\alias{get_fxdf}
\title{Get the data frame generated by a table build function.}
\usage{
get_fxdf(fx.name, all.npos, time, year)
}
\arguments{
\item{fx.name}{A function name from get_fx_names().}

\item{all.npos}{A list of parsed NPO data.}

\item{time}{A string representing the timestamp for the file.}

\item{year}{An integer specifying the tax year.}
}
\value{
None. Writes data to a CSV file.
}
\description{
Helper function that extracts a single table (the table associated with the function name passed as an argument) from a list of multiple tables from multiple nonprofits. It is used by build_tables() to write a batch of the data to file.
}
\examples{
# extract all tables for ten 990 filers
fx.names <- get_fx_names()
timestamp <- format(Sys.time(), "\%b-\%d-\%Y-\%Hh-\%Mm")
# sample of 10 orgs in 2020
i2 <- dplyr::filter( 
      tinyindex, 
      TaxYear == 2020,
      FormType \%in\% c("990","990EZ") )
urls <- i2$URL[1:10]  
# pool data for the given table from npos in the sample
all.npos <- purrr::map( urls, parse_npo, fx.names )
get_fxdf( "BUILD_F9_P01_T00_SUMMARY", all.npos, timestamp, 2020 )
# CREATES FILE: "2020-F9-P01-T00-SUMMARY-Jan-22-2025-15h-13m.csv"
}
