% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{get_missing_xpath_df}
\alias{get_missing_xpath_df}
\title{Process Missing Xpaths Log.}
\usage{
get_missing_xpath_df(path = NULL)
}
\description{
Read a MISSING-XPATHS.txt file and convert the log to a data frame.
}
