% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIAGRAM_TABLE_STRUCTURE.R
\name{get_nd}
\alias{get_nd}
\title{Get Node Tree}
\usage{
get_nd(table.name)
}
\arguments{
\item{table.name}{A character string specifying the table name.}
}
\value{
A data.tree object representing the table's hierarchical structure.
}
\description{
Constructs a data.tree structure for a given table based on its XPaths.
}
\examples{
get_nd("F9-P03-T01-PROGRAMS-OTHER")
}
