% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{get_object_id}
\alias{get_object_id}
\title{Extract Object ID from URL}
\usage{
get_object_id(url)
}
\arguments{
\item{url}{A string containing the full URL to parse.}
}
\value{
A string representing the extracted object ID.
}
\description{
Parses and extracts the unique object ID from a given URL.
}
\examples{
base <- "https://gt990datalake-rawdata.s3.amazonaws.com/EfileData/XmlFiles/"
fn <- "202301529349200315_public.xml"
url <- paste0( base, fn )
id <- get_object_id(url)
# "OID-202301529349200315"
}
