% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb-build-function-adapted.R
\name{get_table_2}
\alias{get_table_2}
\title{get_table_2 (improved from rdb-functions-v2)}
\usage{
get_table_2(doc, group.names, table.name)
}
\description{
collects the information out of the relevant xpath roots
including the subgrouped roots and data as well
}
