% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIAGRAM_TABLE_STRUCTURE.R
\name{get_table_names}
\alias{get_table_names}
\title{Get Table Names}
\usage{
get_table_names(concordance = NULL)
}
\arguments{
\item{concordance}{A data frame or data.table of concordance data. Defaults to NULL, which fetches the concordance.}
}
\value{
A character vector of unique table names.
}
\description{
Retrieves a unique list of table names from the concordance table.
}
\examples{
get_table_names()
}
