% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_FUNCTIONS_RDB.R
\name{get_table_v2}
\alias{get_table_v2}
\title{Extract Table Data from XML}
\usage{
get_table_v2(doc, table.name, table.headers)
}
\arguments{
\item{doc}{An XML document object.}

\item{table.name}{A character string representing the table name.}

\item{table.headers}{A character vector of table header XPaths.}
}
\value{
A data frame containing the extracted table data or NULL if no data is found.
}
\description{
Extracts table data from an XML document based on specified headers.
}
\examples{
table_data <- get_table_v2(doc, "F9-P03-T01-PROGRAMS", headers)
}
