% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb-build-function-adapted.R,
%   R/rdb-functions-v2.R
\name{get_var_map}
\alias{get_var_map}
\title{get_var_map (same as rdb-functions-v2)}
\usage{
get_var_map(table.name)

get_var_map(table.name)
}
\description{
creates a crosswalk between IRS

some description text
}
\details{
some additional details
}
