% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{get_year}
\alias{get_year}
\title{Filter Dataset by Year}
\usage{
get_year(index, year)
}
\arguments{
\item{index}{A data frame containing tax year information.}

\item{year}{The tax year to filter.}
}
\value{
A data frame filtered by the specified year.
}
\description{
Subsets a dataset for a specified tax year.
}
\examples{
filtered_data <- get_year(index, 2020)
}
