% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{log_collapsed_record}
\alias{log_collapsed_record}
\title{Record a change to raw data.}
\usage{
log_collapsed_record(varname, ein = ORG_EIN, year = TAX_YEAR, url = URL)
}
\description{
Make a note when a variable contained within a
one-to-one table is collapsed into a single value.
}
\details{
Some variables in the 1:1 table list occassionally contain
more than one value. For example, most nonprofits are registered in
a single state but a small proportion are registered in several.
These values are collapsed into a single value, for example
state = MN, OH, and NY becomes a single string \verb{\{MN\},\{OH\},\{NY\}}.
}
