% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{log_failed_url}
\alias{log_failed_url}
\title{Log a failed URL.}
\usage{
log_failed_url(URL)
}
\description{
Make a note when a url fails to load or is invalid.
}
\details{
Poor internet connections, long server latency, or invalid
urls, or corrupted XML structure will result in an XML file failing to load.
}
