% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{log_missing_xpaths}
\alias{log_missing_xpaths}
\title{Log missing xpaths.}
\usage{
log_missing_xpaths(doc, url)
}
\description{
Create a log of which xpaths are currently missing from the concordance file.
}
