% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{namedList}
\alias{namedList}
\title{Create a Named List}
\usage{
namedList(...)
}
\arguments{
\item{...}{Named arguments to include in the list.}
}
\value{
A named list.
}
\description{
Constructs a named list from the provided arguments.
}
\examples{
my_list <- namedList(a = 1, b = 2)
}
