% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-one-to-one-tables.R
\name{parsapply_tables}
\alias{parsapply_tables}
\title{Passing arguments to parSapply}
\usage{
parsapply_tables(index.group)
}
\description{
Pass arguments to parallel sapply table function.
}
\details{
Helper function to send variables to the build_tables function in parSapply framework.
}
