% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIAGRAM_TABLE_STRUCTURE.R
\name{plot_table_str}
\alias{plot_table_str}
\title{Plot Table Structure}
\usage{
plot_table_str(table.name)
}
\arguments{
\item{table.name}{A character string specifying the table name.}
}
\value{
None. Displays a plot of the table structure.
}
\description{
Plots the hierarchical structure of a table using data.tree visualization.
}
\examples{
plot_table_str("F9-P03-T01-PROGRAMS-OTHER")
}
