% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FULL_DATABASE.R
\name{prep_index}
\alias{prep_index}
\title{Apply Specified Filters to the Build Index}
\usage{
prep_index(years = NULL, index = NULL, form.type = c("990", "990EZ"))
}
\arguments{
\item{index}{A Data Commons index from build_index(), get_current_index_full(), or get_current_index_batch().}

\item{year}{The desired years of the database to create.}
}
\value{
A filtered index file.
}
\description{
Apply Specified Filters to the Build Index
}
\examples{
index <- get_current_index_batch()
index <- prep_index( index, years=2023 )
}
