% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIAGRAM_TABLE_STRUCTURE.R
\name{print_table_str}
\alias{print_table_str}
\title{Print Table Structure}
\usage{
print_table_str(table.name)
}
\arguments{
\item{table.name}{A character string specifying the table name.}
}
\value{
None. Prints the table structure to the console.
}
\description{
Prints the hierarchical structure of a table.
}
\examples{
print_table_str("F9-P03-T01-PROGRAMS-OTHER")
}
