% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_FUNCTIONS.R
\name{process_batch}
\alias{process_batch}
\title{Process and Build Batches in Parallel}
\usage{
process_batch(batch, year, fx.names = NULL)
}
\arguments{
\item{year}{The tax year associated with the data.}

\item{fx.names}{A vector of function names for processing tables.}

\item{batch.ids}{A list of batch IDs to process.}

\item{batch.list}{The complete batch list containing URLs.}
}
\value{
A vector of completed batch IDs.
}
\description{
Processes batch IDs using \code{furrr::future_map()}
}
