% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODE_CHUNKS.R
\docType{data}
\name{RDB.KEYS}
\alias{RDB.KEYS}
\title{RDB KEYS fields.}
\format{
An object of class \code{character} of length 10.
}
\usage{
RDB.KEYS
}
\description{
List of relational database fields that are present on all tables and are used for filtering observations and merges.
}
\details{
A character vector listing fields used as keys, mainly for internal package use.
}
\keyword{datasets}
