% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FULL_DATABASE.R
\name{remove_batch}
\alias{remove_batch}
\title{Remove Batch of URLs from Queue}
\usage{
remove_batch(x)
}
\arguments{
\item{x}{Batch name ("g1","g2",...) created from split_index()}
}
\description{
Extracts and writes table data from URLs.
}
\examples{
groups <- split_index( tinyindex )
L <- readRDS("BATCHFILE.RDS")
names(L)
remove_batch( "g3" )
L <- readRDS("BATCHFILE.RDS")
names(L)
}
