% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb-build-function-adapted.R,
%   R/rdb-functions-v2.R
\name{remove_count}
\alias{remove_count}
\title{remove_count (same as rdb-functions-v2)}
\usage{
remove_count(x)

remove_count(x)
}
\description{
gets rid of extraneous numbering and details in
potential group names

some description text
}
\details{
uses regex to do this

some additional details
}
