% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FULL_DATABASE.R
\name{remove_groups}
\alias{remove_groups}
\title{Remove Groups of URLs from Queue}
\usage{
remove_groups(groups)
}
\arguments{
\item{x}{Batch name ("g1","g2",...) created from split_index()}
}
\description{
Extracts completed batches from BATCHFILE.
}
\examples{
groups <- split_index( tinyindex, group.size=25 )
L <- readRDS("BATCHFILE.RDS")
names(L)
remove_groups( c("G2{25}","G3{25}") )
L <- readRDS("BATCHFILE.RDS")
names(L)
}
