% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FULL_DATABASE.R
\name{resume_build_database}
\alias{resume_build_database}
\title{Resume build_database() if Interrupted.}
\usage{
resume_build_database(years = NULL, index = NULL)
}
\arguments{
\item{years}{A vector of years remaining in the build. If NULL, all index years will be attempted.}

\item{index}{Use the same index object that was originally passed to build_database().
If left NULL it loads the version that is saved when build_database() is first called (build-index.rds).}
}
\description{
If you need to pause build_database() before
it is complete or your computer restarts you can resume
the process from the stop point.
}
