% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{retrieve_xml}
\alias{retrieve_xml}
\title{Extract Text from an XML Node}
\usage{
retrieve_xml(doc, TEMP_VAR)
}
\arguments{
\item{doc}{An XML document object.}

\item{TEMP_VAR}{The XPath expression specifying the nodes to retrieve.}
}
\value{
A character vector containing the extracted text or NA if no text is found.
}
\description{
Retrieves text from specific nodes within an XML document.
}
\examples{
xml_data <- retrieve_xml(doc, "//TagName")
}
