% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_FUNCTIONS.R
\name{send_batch}
\alias{send_batch}
\title{Error handling and messaging for parsupply_tables}
\usage{
send_batch(batch.ids, cl)
}
\description{
Pass a group of batch IDs to parallel sapply table function.
}
\details{
Logging and resource management for a subgroup of batches to prevent BATCHFILE read write conflicts.
}
