% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-one-to-one-tables.R
\name{split_index}
\alias{split_index}
\title{Split index into groups}
\usage{
split_index(index, group.size = 1000)
}
\description{
Split the index into \code{group.size} in order
to facilitate batch processing of files.
}
\details{
some additional details
}
