% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FULL_DATABASE.R
\name{split_into_batches}
\alias{split_into_batches}
\title{Split Groups into Batches}
\usage{
split_into_batches(glist, numcores)
}
\arguments{
\item{glist}{A list of grouped items, typically from \code{split_into_groups()}.}

\item{numcores}{An integer specifying the number of batches (e.g., number of CPU cores for parallel processing).}
}
\value{
A named list of batches, each containing a subset of groups.
}
\description{
Organizes a list of groups into batches for parallel processing.
}
