% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FULL_DATABASE.R
\name{split_into_groups}
\alias{split_into_groups}
\title{Split Vector into Groups}
\usage{
split_into_groups(x, G = 25)
}
\arguments{
\item{x}{A vector of items to be split.}

\item{G}{An integer specifying the maximum size of each group (default: 25).}
}
\value{
A named list where each element is a vector of length up to \code{G}.
}
\description{
Divides a vector into smaller groups of a specified size.
}
