% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{standardize_boole}
\alias{standardize_boole}
\title{Standardize Boolean Inputs}
\usage{
standardize_boole(x)
}
\arguments{
\item{x}{A vector of values to standardize.}
}
\value{
A logical vector indicating TRUE/FALSE values.
}
\description{
Converts various representations of true/false into a standardized boolean format.
}
\examples{
standardize_boole(c("YES", "NO", NA))
}
