% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{test_build}
\alias{test_build}
\title{Package Testing Function}
\usage{
test_build(path = ".", spare.cores = 1)
}
\arguments{
\item{path}{The folder in which the test build will occur (if no value provided it uses the default folder).}
}
\description{
Build a small database to test the package.
}
\examples{
test_build()
}
