% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tinyindex}
\alias{tinyindex}
\title{Small AWS efile index created for demo and testing purposes.}
\format{
A data frame with 10000 rows and 10 variables:
\describe{
\item{OrganizationName}{Name of the nonprofit filing the 990 form}
\item{EIN}{The nonprofit Employer Identification Number, i.e. the unique org ID}
\item{FormType}{Type of tax form used (990,990EZ,990PF)}
\item{TaxPeriod}{Due date of the tax filing in format YYYY-MM}
\item{TaxYear}{The calendar year covered by the tax filing (TaxPeriod-1) in format YYYY}
\item{URL}{Online address of the XML file containing efile data for the organization}
\item{DLN}{IRS Document Locator Number - internal ID used to track efilings}
\item{SubmittedOn"}{Date the filing was submitted}
\item{ObjectId}{Internal IRS field}
\item{LastUpdated}{Internal IRS field}
}
}
\source{
\url{https://registry.opendata.aws/irs990/}
}
\description{
A dataset reporting the meta-data for 10,000 nonprofit efile returns.
The index gives the URL (web address) of each XML file stores in the AWS instance
and is the basis for an iterative data build in this package.
}
