% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{update_concordance}
\alias{update_concordance}
\title{Update concordance Data Object}
\usage{
update_concordance()
}
\description{
A package maintainance function that updates the concordance
data object data(concordance) that is used when the github version can't be retrieved.
}
\examples{
update_concordance()  # run inside main package folder
}
