% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{update_tinyindex}
\alias{update_tinyindex}
\title{Update tinyindex Data Object}
\usage{
update_tinyindex()
}
\description{
A package maintainance function that updates the tinyindex
data index that is used for testing and development. A fresh index is
downloaded from the Data Commons and a new sample saved with usethis::use_data().
}
\examples{
update_tinyindex()  # run inside main package folder
}
