% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{update_xpaths}
\alias{update_xpaths}
\title{Update xpaths Data Object}
\usage{
update_xpaths()
}
\description{
A package maintainance function that updates the xpaths
object used for logging missing xpaths using the usethis package.
}
\examples{
update_xpaths()  # run inside main package folder
}
