% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_INDEX.R
\name{valid_url}
\alias{valid_url}
\title{Check if URL is valid}
\usage{
valid_url(URL, t = 2)
}
\arguments{
\item{URL}{Any URL you wish to test.}
}
\value{
Returns a value of TRUE or FALSE.
}
\description{
Check whether a URL is valid.
}
\details{
Pings a given URL to see if it is active or exists.
}
\examples{
\dontrun{
# TRUE
valid_url( "https://nccs-efile.s3.us-east-1.amazonaws.com/index/data-commons-index-file-2020.csv" )
# FALSE
valid_url( "https://nccs-efile.s3.us-east-1.amazonaws.com/index/data-commons-index-file-2090.csv" )
}
}
