% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb-build-function-adapted.R,
%   R/rdb-functions-v2.R
\name{validate_group_names}
\alias{validate_group_names}
\title{validate_group_names function (same as rdb-functions-v2)}
\usage{
validate_group_names(nd, table.name)

validate_group_names(nd, table.name)
}
\description{
Validating group names from a nodeset (nd) and
a specific table (table.name)

some description text
}
\details{
some additional details

some additional details
}
