% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{zeroALL}
\alias{zeroALL}
\title{General Zero Replacement}
\usage{
zeroALL(var)
}
\arguments{
\item{var}{A variable to evaluate.}
}
\value{
The original variable or "0" if it is missing/NA.
}
\description{
Replaces missing or NA values with "0" for all data types.
}
