% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-full-database.R
\name{bind_data}
\alias{bind_data}
\title{Combine table segments.}
\usage{
bind_data(years)
}
\description{
Combine all of the raw table files created during parallel processing.
}
\details{
To be added...
}
