% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-full-database.R
\name{build_database}
\alias{build_database}
\title{Build a complete database}
\usage{
build_database(index = NULL, years = NULL)
}
\arguments{
\item{index}{An index file created from \code{build_index()}.}

\item{years}{The TaxYears to include in the build.}
}
\value{
Saves log files and database tables to your computer.
}
\description{
Build a database of IRS 990 efilers available on AWS
}
\details{
To be added...
}
\examples{
\dontrun{
index <- build_index( )
build_database( index, years=2015:2017 )
}
}
