% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_FUNCTIONS.R
\name{build_tables_parallel}
\alias{build_tables_parallel}
\title{Parallel Build of Tables}
\usage{
build_tables_parallel(batch.list, year, fx.names = NULL)
}
\arguments{
\item{year}{The tax year associated with the data.}

\item{fx.names}{A vector of function names for processing tables.}

\item{groups}{A list of URL groups to process.}
}
\value{
A vector of failed URLs.
}
\description{
Builds tables in parallel using multiple cores.
}
\examples{
failed_urls <- build_tables_parallel(groups, year = 2023)
}
