% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_FUNCTIONS_RDB.R
\name{collapse_nodes}
\alias{collapse_nodes}
\title{Collapse Nodes}
\usage{
collapse_nodes(lc.xpaths)
}
\arguments{
\item{lc.xpaths}{A character vector of XPath prefixes.}
}
\value{
A character vector of consolidated XPath group names.
}
\description{
Consolidates and generates group names for given XPaths.
}
\details{
This function uses the concordance table to identify and collapse nodes for the provided XPaths.
}
\examples{
collapsed_nodes <- collapse_nodes(c("IRS990", "IRS990EZ"))
}
