% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_FUNCTIONS.R
\name{get_fx_names}
\alias{get_fx_names}
\title{Get Function Names}
\usage{
get_fx_names(table.names = NULL)
}
\arguments{
\item{table.names}{A vector of table names.}
}
\value{
A vector of function names.
}
\description{
Retrieves a list of function names based on table names and exclusion criteria.
}
\details{
Defaults to all tables in the concordance minus T99 table
(supplementary information sections). The BUILD_SCHEDULE_TABLE()
function is not defined in the concordance, but is added to the list.
}
\examples{
table_names <- get_table_names( exclude="T99" )
fx_names <- get_fx_names( table_names )
}
