% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIAGRAM_TABLE_STRUCTURE.R, R/TABLE_FUNCTIONS.R
\name{get_table_names}
\alias{get_table_names}
\title{Get Table Names}
\usage{
get_table_names(exclude = c("T99"))

get_table_names(exclude = c("T99"))
}
\arguments{
\item{exclude}{A vector of substrings to exclude from table names.}

\item{concordance}{A data frame or data.table of concordance data. Defaults to NULL, which fetches the concordance.}
}
\value{
A character vector of unique table names.

A vector of table names.
}
\description{
Retrieves a unique list of table names from the concordance table.

Retrieves the list of unique table names defined in the Concordance rdb_table field.
}
\examples{
get_table_names()
table_names <- get_table_names()
}
