% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_FUNCTIONS.R
\name{parse_npo}
\alias{parse_npo}
\title{Parse NPO Data}
\usage{
parse_npo(url, fx.names, logXP = TRUE)
}
\arguments{
\item{url}{A string containing the URL of the XML file.}

\item{fx.names}{A list of functions to apply to the XML document.}
}
\value{
A named list with parsed data or a failure indicator if the URL is inaccessible.
}
\description{
Parses XML data from a given URL and applies specified processing functions.
}
\examples{
npo_data <- parse_npo(url, fx.names)
}
