% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb-build-function-adapted.R,
%   R/rdb-functions-v2.R
\name{re_name}
\alias{re_name}
\title{re_name (same as rdb-functions-v2)}
\usage{
re_name(df, v.map)

re_name(df, v.map)
}
\description{
converts IRS names to concordance names in a dataframe

some description text
}
\details{
some additional details
}
