% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FULL_DATABASE.R
\name{split_index}
\alias{split_index}
\title{Split Index into Groups}
\usage{
split_index(year, index, group.size = 1000)
}
\arguments{
\item{index}{A data frame containing a column \code{URL} with URLs to be split.}

\item{group.size}{An integer specifying the maximum size of each group.}
}
\value{
A list of URL groups.
}
\description{
Divides a vector of URLs into smaller groups of a specified size.
}
\examples{
groups <- split_index(index, group.size = 1000)
}
