% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb-functions-v2.R
\name{validate_group_names}
\alias{validate_group_names}
\title{some title text}
\usage{
validate_group_names(nd, original.xpaths)
}
\description{
some description text
}
\details{
some additional details
}
